Fix Matches method for FD values.
This would cause most attempted matches on FD values to fail,
and could also potentially cause matches to succeed when they
should not.

+++ vtkDICOMValue.cxx
@@ -2519,7 +2519,7 @@ bool vtkDICOMValue::Matches(const vtkDICOMValue& value) const
   else if (type == VTK_DOUBLE)
     {
     // Match if any value matches
-    match = ValueT<float>::CompareEach(value.V, this->V);
+    match = ValueT<double>::CompareEach(value.V, this->V);
     }
 
   return match;
